Moved ImporterErrors tests to test_importer.py and refactor.
